So far gsMatrixOp only stores a reference to the provided matrix and requires the user to take care of storing the matrix. So, a corresponding warining was added.
Moreover, a specialication of the template was added, which takes a shared pointer to the matrix and stores that. This is of particular interest if a function is supposed to set up a matrix, to bulid a gsMatrixOp from it and to return that to the user.

The same was done for gsTransposedMatrixOp. For consitency, also gsSolverOp was updated to take smart pointers.

I hope, this works on all compilers.



git-svn-id: https://ricamsvn.ricam.oeaw.ac.at/gismo/stable@7577 e128ff1e-fa79-4151-a8cb-522299a2c450