gsIterativeSolver.h: it might be good to swhich of the new feature.
moreover, there seemed to be a bug in solve(). in the very first round,
the algorithm reads a m_error which is not (properly) initialized.

i have removed the push_back to the error history in the initialization
because this violates the assumption that the vector size should be as
large as m_num_iter.



git-svn-id: https://ricamsvn.ricam.oeaw.ac.at/gismo/stable@8079 e128ff1e-fa79-4151-a8cb-522299a2c450