Consistency change for gsFileData.h
The behavior of gsFileData should be similar for the case that the file cannot be found and the case that the file is invalid.

Moreover, it might be of interest to return a bool on gsFileData::read.

So, finally the consumer can (in all cases) just call
  if (fd.hasAny< ObjectTypeOfInterest >())
to check all cases (file not existing, invalid file, file without that object).