the partition function is a total overkill if one only needs the number of free uncoupled dofs
moreover, the function partition() is dubious since it does not ask for the additional information which is necessary to setup a dof mapper (like the interface strategy and the Dirichlet strategy). After setting up the dof mapper,
the function paritition() only accesses the dofmapper, so it might actually be a member of dofmapper instead.