IETI example for Stokes (#708)
Add a new example that shows how Expression Assembler + IETI framework can be used to solve systems of PDEs.

A corresponding geometry and a comprehensive documentation has been added. Moreover, this improves IETI plot functionality by directly constructing the field to be written out for Paraview. There is no need to have to setup a new assembler just for this purpose since the IETI mapper has all the data required to incorporate the Dirichlet conditions.